requirements.txt
flask==3.0.3
flask-login==0.6.3
flask-sqlalchemy==3.1.1
werkzeug==3.0.3
requests==2.31.0
python-dotenv==1.0.1
bcrypt==4.1.3
email-validator==2.1.1
gunicorn==21.2.0
